requirements.txt
# The runtime dependencies of the framework, as if via `pip install kopf`.
-e .

# Everything needed to develop (test, debug) the framework.
coverage
pytest-aiohttp
pytest-asyncio
pytest-mock
pytest-cov
pytest>=6.0.0
aresponses
asynctest
async-timeout
freezegun
codecov
coveralls
flake8
mypy==0.800
isort>=5.5.0
